Backport support for the Sony Sixaxis and DualShock 4 controllers.
This backports support for force-feedback on both the Sixaxis and
DualShock 4, full LED support for both controllers, battery reporting for
both controllers and touchpad support on the DS4.  All features are
functional under both USB and Bluetooth.

Additionally, this adds a workaround to stop the Sixaxis from blinking
after the PS button is pushed when connected via USB and it uses the same
trick as the xpad driver to set the default LED values using the joystick
device number.

Note that the DualShock 4 needs BlueZ 5.14 or newer to pair correctly.

Signed-off-by: Frank Praznik <frank.praznik@oh.rr.com>