Workaround for sending reports to the Sixaxis and DualShock 4.
The Sixaxis and DualShock 4 want output reports sent on the control
channel when running on Bluetooth, so a special case is needed for them.

This is a bit of an ugly hack, but it gets the job done without needing to
backport the rather extensive HID system changes in newer kernels that are
needed to make these controllers function via Bluetooth.

Signed-off-by: Frank Praznik <frank.praznik@oh.rr.com>