xpad: fix xbox wireless gamepad LED initialization
Immediately after becoming present, we send the LED change
request to the pad.  There is then a race where the input
device may be opened and closed before the request's urb
is actually transmitted, in which case the LED change is
just dropped because of the usb_kill_urb().

This was actually happening almost all the time for me.

We really still want any outstanding output state changes to
get through, so just don't do usb_kill_urb() on the wireless
xbox360 device.

(This issue is perfectly likely to also affect other pad types,
but I've been conservative with the change.  It may also
help with a similar issue with dropped rumble state change
immediately before a close - unverified.)