Input: xpad: handle "present" and "gone" correctly
From: "Pierre-Loup A. Griffais"

Handle the "a new device is present" message properly by dynamically
creating the input device at this point in time.  This requires a
workqueue as we are in interrupt context when we learn about this.

Also properly disconnect any devices that we are told are removed.